moved listParentSkillsAction to a newly made SkillController and also adjusted base.html
